Updated deferred pipeline to no longer be full pbr or zero pbr. Can use a mix of both lighting and material types

+++ CustomGameEngine/Shaders/g_buffer.frag
@@ -3,7 +3,7 @@ layout (location = 0) out vec3 gPosition;
 layout (location = 1) out vec3 gNormal;
 layout (location = 2) out vec3 gAlbedo;
 layout (location = 3) out vec4 gSpecular;
-layout (location = 5) out uint gPBRFLAG;
+layout (location = 5) out float gPBRFLAG;
 
 in VIEW_DATA {
     flat vec3 TangentViewPos;
@@ -127,5 +127,5 @@ void main() {
     gSpecular.rgb = Specular;
     gSpecular.a = material.SHININESS;
 
-    gPBRFLAG = uint(0);
+    gPBRFLAG = 0.0;
 }
